New structure for files controllers 📦
Divided the functions into more precise controllers. So we don't have long complex files.
