+++ app/views/files/browse/_file_browse.html.erb
