--- app/views/files/browse/destroy_folder.json.jbuilder
@@ -1,2 +1,0 @@
-json.success @success
-json.message @success ? 'deleted all files in folder' : 'could not delete all files'
