+++ app/views/files/files/new.html.erb
@@ -10,7 +10,7 @@
 </script>
 
 <script type="text/ng-template" id="fileBrowse">
-  <%= render 'files/files/file_browse' %>
+  <%= render 'files/browse/file_browse' %>
 </script>
 
 <script type="text/ng-template" id="tagCreate">
