+++ app/views/files/upload/upload_from_base64.json.jbuilder
